Split firm economics stages
